+++ triangle_java/Scalene.java
@@ -2,12 +2,12 @@ package module;
 
 public class Scalene extends Triangle {
 
-	public Scalene(float a, float b, float c) {
+	public Scalene(int a, int b, int c) {
 		super(a, b, c);
 	}
 
 
-	public static Boolean isScalene(float a, float b, float c) {
+	public static Boolean isScalene(int a, int b, int c) {
 		if (a != b && b != c && c != a) {
 			System.out.println("Je suis Scalène ");
 			return true;
